完成后端Web实战

+++ tlias-web-management/src/main/java/io/github/csclly/service/impl/DeptServiceImpl.java
@@ -1,6 +1,8 @@
 package io.github.csclly.service.impl;
 
+import io.github.csclly.exception.BusinessException;
 import io.github.csclly.mapper.DeptMapper;
+import io.github.csclly.mapper.EmpMapper;
 import io.github.csclly.poji.Dept;
 import io.github.csclly.service.DeptService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,6 +15,8 @@ import java.util.List;
 public class DeptServiceImpl implements DeptService {
     @Autowired
     private DeptMapper deptMapper;
+    @Autowired
+    private EmpMapper empMapper;
     @Override
     public List<Dept> findAll() {
         return deptMapper.findAll();
@@ -20,6 +24,11 @@ public class DeptServiceImpl implements DeptService {
 
     @Override
     public void deleteById(Integer id) {
+        Integer count = empMapper.countById(id);
+        if (count > 0) {
+            throw new BusinessException("对不起，当前部门下有员工，不能直接删除");
+        }
+
         deptMapper.deleteById(id);
     }
